restapi/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
python-dotenv==0.12.0
Flask-SQLAlchemy==2.4.1
